% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot.R
\name{lineplot}
\alias{lineplot}
\title{Generate a Line Plot}
\usage{
lineplot(data, cols)
}
\arguments{
\item{data}{A dataframe with rows of data and columns of variables. The
The dataframe should contain the columns specified in \code{cols}.}

\item{cols}{A vector of column names in \code{data} to be plotted.
These columns should contain categorical or discrete numerical values
for which counts will be plotted.}
}
\value{
A \code{ggplot} object representing a line plot of counts of unique
values for the specified columns in \code{data}. Each line is colored
differently to distinguish between the variables represented by the
columns in \code{cols}.
}
\description{
This function takes a dataframe and a list of column names as input and
generates a line plot. The plot displays the counts of unique values in
the specified columns. Each line in the plot represents one of the
specified columns, differentiated by color.
}
\examples{
data <- data.frame(
  Names = c("Alice", "Bob", "Charlie"),
  Ages = c(25, 30, 22),
  Height = c(21, 33, 21))

cols <- c("Ages", "Height")
lineplot(data, cols)
}
